refactor: reorganize command modules
