+++ src/commands/compare/compareCommands.ts
