+++ src/commands/image/imageCommands.ts
