+++ src/commands/text/textEditorCommands.ts
