+++ src/commands/workspace/tests/connectionTests.ts
